+++ src/help/zipObject.js
@@ -1,7 +1,17 @@
 function zipObject ( keys, values ) {   // (string[], string[] ) -> object
     // *** Conect two arrays in a single object
         return keys.reduce ( (res,k,i)  => {
-                                let  val = Number (values[i]) ? parseInt(values[i]) : values[i];
+                                let 
+                                      income = values[i]
+                                    , val
+                                    ;
+                                switch ( typeof income ) {
+                                            case 'string' :
+                                                    val = Number(income) || income
+                                                    break
+                                            default:
+                                                    val = income
+                                    }
                                 res[k] = val
                                 return res
                     }, {} )